refactor: Move theme related components into folder
